+++ start-signingsuite.ps1
@@ -22,7 +22,7 @@
 
 .NOTES
     ScriptName : start-signingsuite.ps1
-    Version    : 2026.09.15.0004
+    Version    : 2026.09.15.0005
 #>
 [CmdletBinding()]
 param(